refactor(mcp-server): relocate DaemonBackend transport out of src/app
daemon-backend, api-client, upload-files, ws-text-message, and
browser-tracing move from src/app/lib (deletion-slated in Stage 5) to
src/shared, unblocking apps/web from importing DaemonBackend via a new
./daemon-backend package subpath export. src/app/lib keeps thin
re-export shims so existing imports (useWhiteboardSync.ts,
useBranches.ts, main.tsx, canvas-backend.ts) and their tests are
unchanged. api-client's window/RequestInfo usage is rewritten to a
structurally-typed globalThis read and a Request|string|URL union so
the module compiles under both tsconfig.server.json (no DOM lib) and
apps/web's DOM-enabled tsconfig.
