+++ packages/mcp-server/src/app/lib/ws-text-message.ts
@@ -1,24 +1,4 @@
-import { type ServerTextMessage, serverTextMessageSchema } from '../../shared/ws-messages.js'
-
-export function parseServerTextMessage(
-  raw: string,
-  warn: (...args: unknown[]) => void = console.warn,
-): ServerTextMessage | null {
-  let parsed: unknown
-  try {
-    parsed = JSON.parse(raw)
-  } catch {
-    warn('[whiteboard] ignored invalid server text message:', 'malformed JSON', raw)
-    return null
-  }
-  const result = serverTextMessageSchema.safeParse(parsed)
-  if (!result.success) {
-    warn(
-      '[whiteboard] ignored invalid server text message:',
-      result.error.issues[0]?.message ?? 'schema mismatch',
-      parsed,
-    )
-    return null
-  }
-  return result.data
-}
+// Relocated to src/shared/ws-text-message.ts as part of the daemon-backend
+// transport move. Re-exported here so existing src/app imports keep working
+// unchanged.
+export { parseServerTextMessage } from '../../shared/ws-text-message.js'
